Initial commit
Added a basic UITableView within a custom  like subclass of UIViewController for debugging. This used the bottom sheet in the same way that  does. I don't think that BottomSheet is the bug. This lead me to believe setting the bottomSheetPreferredHeight in ActionSheet, we are setting it to the wrong value